Principal-main/desafio5/data-service/requirements 2.txt
flask
pyjwt
requests
